[CLI] - Create Workflow Command
Currently the creation of a workflow is via 'tnctl workflow create'; given we now have a 'tnctl create' verb now, feels best to move the command
