+++ pkg/cmd/tnctl/tnctl.go
@@ -40,7 +40,6 @@ import (
 	"github.com/appvia/terranetes-controller/pkg/cmd/tnctl/retry"
 	"github.com/appvia/terranetes-controller/pkg/cmd/tnctl/search"
 	"github.com/appvia/terranetes-controller/pkg/cmd/tnctl/state"
-	"github.com/appvia/terranetes-controller/pkg/cmd/tnctl/workflow"
 	"github.com/appvia/terranetes-controller/pkg/version"
 )
 
@@ -81,7 +80,6 @@ func New(factory cmd.Factory) *cobra.Command {
 		search.NewCommand(factory),
 		convert.NewCommand(factory),
 		create.NewCommand(factory),
-		workflow.NewCommand(factory),
 		describe.NewCommand(factory),
 		generate.NewCommand(factory),
 		state.NewCommand(factory),
